+++ ui/apps/platform/cypress/constants/apiEndpoints.js
@@ -53,7 +53,7 @@ export const risks = {
         'v1/deploymentswithprocessinfo?query=&pagination.offset=0&pagination.limit=50&pagination.sortOption.field=Priority&pagination.sortOption.reversed=false',
     deploymentsCount: 'v1/deploymentscount?query=',
     getDeployment: 'v1/deployments/*',
-    getDeploymentWithRisk: 'v1/deploymentswithrisk/*',
+    fetchDeploymentWithRisk: 'v1/deploymentswithrisk/*',
     graphqlOps: {
         autocomplete: 'autocomplete',
         getProcessTags: 'getProcessTags',