+++ ui/apps/platform/cypress/integration/risk/podTimeline.test.js
@@ -8,7 +8,7 @@ import checkFeatureFlag from '../../helpers/features';
 function setRoutes() {
     cy.server();
     cy.route('GET', api.risks.riskyDeployments).as('deployments');
-    cy.route('GET', api.risks.getDeploymentWithRisk).as('getDeployment');
+    cy.route('GET', api.risks.fetchDeploymentWithRisk).as('getDeployment');
     cy.route('POST', api.graphql(api.risks.graphqlOps.getDeploymentEventTimeline)).as(
         'getDeploymentEventTimeline'
     );