+++ ui/apps/platform/cypress/integration/risk/processtags.test.js
@@ -11,7 +11,7 @@ import withAuth from '../../helpers/basicAuth';
 function setRoutes() {
     cy.server();
     cy.route('GET', api.risks.riskyDeployments).as('deployments');
-    cy.route('GET', api.risks.getDeploymentWithRisk).as('getDeployment');
+    cy.route('GET', api.risks.fetchDeploymentWithRisk).as('getDeployment');
     cy.route('POST', api.graphql(api.risks.graphqlOps.getProcessTags)).as('getTags');
     cy.route('POST', api.graphql(api.risks.graphqlOps.autocomplete)).as('tagsAutocomplete');
 }